+++ include/mem/std.h
@@ -3,6 +3,6 @@
 
 #include <stddef.h>
 
-void *std_alloc(size_t size, unsigned int alignment);
+void *std_alloc(unsigned int size, unsigned int alignment);
 
 unsigned char std_free(void **ptr);
